backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
torch
torchvision
Pillow
